+++ Sources/RouterBytes/APIRouter.swift
@@ -61,7 +61,7 @@ public struct RetryOptions: OptionSet, Sendable {
 public protocol APIRouter<RequestBody>: Sendable {
     associatedtype Response: Sendable = Void
     associatedtype HeaderResponse: Sendable = Void
-    associatedtype AuthorizationType = RouterBytes.AuthorizationType
+    associatedtype AuthorizationType = RouterBytes.BearerAuthorizationType
     associatedtype RequestBody: Sendable = Void
 
     // Properties to be specified within the project APIRouter protocol