+++ Sources/RouterBytesAuthentication/APITokenAuthorizationType.swift
@@ -10,42 +10,7 @@ import RouterBytes
 
 @available(macOS 10.15.0, *)
 public protocol APITokenAuthorizationType {
-    func authorizedRequest(request: HTTPRequest, with apiToken: some APITokenType) async throws -> HTTPRequest
-    func authorizedRequest(request: HTTPRequest, with provider: some APITokenProvider) async throws -> HTTPRequest
-}
-
-@available(macOS 10.15.0, *)
-extension RouterBytes.AuthorizationType: APITokenAuthorizationType {
-    public func authorizedRequest(request: HTTPRequest, with apiToken: some APITokenType) -> HTTPRequest {
-        switch self {
-        case let .bearer(tokenType):
-            let token: String
-
-            switch tokenType {
-            case .accessToken:
-                token = apiToken.accessToken.description
-            case .refreshToken:
-                if let apiToken = apiToken as? (any RefreshableAPITokenType) {
-                    token = apiToken.refreshToken.description
-                } else {
-                    assertionFailure("APIToken is not refreshable") // TODO: Find out a typesafe way if possible
-                    token = ""
-                }
-            }
-
-            return request.withBearerToken(token)
-        case .none:
-            return request
-        }
-    }
-
-    public func authorizedRequest(request: HTTPRequest, with provider: some APITokenProvider) async throws -> HTTPRequest {
-        switch self {
-        case .bearer:
-            let apiToken = try await provider.apiToken
-            return authorizedRequest(request: request, with: apiToken)
-        case .none:
-            return request
-        }
-    }
+    associatedtype APIToken: APITokenType
+    func authorizedRequest(request: HTTPRequest, with apiToken: APIToken) async throws -> HTTPRequest
+    func authorizedRequest(request: HTTPRequest, with provider: some APITokenProvider<APIToken>) async throws -> HTTPRequest
 }